Configure DB & Redis variables

+++ database/migrations/2014_10_12_100000_create_password_reset_tokens_table.php
@@ -8,21 +8,29 @@ return new class extends Migration
 {
     /**
      * Run the migrations.
+     * This method handles the creation of the table.
      */
     public function up(): void
     {
-        Schema::create('password_reset_tokens', function (Blueprint $table) {
-            $table->string('email')->primary();
-            $table->string('token');
-            $table->timestamp('created_at')->nullable();
-        });
+        if (!Schema::hasTable('password_reset_tokens')) {
+            Schema::create('password_reset_tokens', function (Blueprint $table) {
+                $table->string('email')->primary();
+                $table->string('token');
+                $table->timestamp('created_at')->nullable();
+            });
+        }
     }
 
     /**
      * Reverse the migrations.
+     * This method handles the removal of the table.
      */
     public function down(): void
     {
-        Schema::dropIfExists('password_reset_tokens');
+        if (Schema::hasTable('password_reset_tokens')) {
+            Schema::dropIfExists('password_reset_tokens');
+        }
     }
 };
+
+
